requirements.txt
Flask
Flask-Session
Flask-SQLAlchemy
Flask-Migrate
Python-Decouple
Marshmallow
PyJWT
Bcrypt
Pymongo
Flask_cors